fix: auth error
