+++ app/api/auth/[...nextauth]/route.ts
@@ -1,15 +1,6 @@
+import { authOptions } from "@/lib/auth";
 import NextAuth from "next-auth";
-import type { NextAuthOptions } from "next-auth";
-import GitHubProvider from "next-auth/providers/github";
 
-export const authOptions: NextAuthOptions = {
-  providers: [
-    GitHubProvider({
-      clientId: process.env.GITHUB_ID!,
-      clientSecret: process.env.GITHUB_SECRET!,
-    }),
-  ],
-};
+const handler = NextAuth(authOptions);
 
-export const handler = NextAuth(authOptions);
 export { handler as GET, handler as POST };
